+++ src/app/providers/index.ts
@@ -1,1 +1,1 @@
-export { QueryProvider } from './QueryProvider'
+export { AppProviders } from './AppProviders'